Rename data-view to fencilla-view
